// src/generated/types/globalTypes.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

//==============================================================
// START Enums and Input Objects
//==============================================================

export interface UserInput {
  uuid?: string | null;
  email?: string | null;
  username?: string | null;
  password?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  avatar?: string | null;
  avatarSource?: string | null;
  avatarX?: number | null;
  avatarY?: number | null;
  avatarScale?: number | null;
}

//==============================================================
// END Enums and Input Objects
//==============================================================